+++ demo/src/test/java/global/eska/ddk/keygen/demo/DemoApplicationTests.java
@@ -11,7 +11,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringRunner;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.Arrays;
 
 import static global.eska.ddk.keygen.utils.Etalon.*;
@@ -53,7 +53,7 @@ public class DemoApplicationTests {
         assertArrayEquals(PRIVATE_KEY_BYTES, keyPair.getSecretKey());
         assertEquals("PUBLIC_KEY_HEX", PUBLIC_KEY_HEX, keyPair.getPublicKeyHex());
 
-        BigDecimal address = accountCreator.getAddressByPublicKey(keyPair.getPublicKey());
+        BigInteger address = accountCreator.getAddressByPublicKey(keyPair.getPublicKey());
         assertEquals("ADDRESS", ADDRESS, address);
 
         log.info("{}", PASSPHRASE);
