+++ lib/src/test/java/global/eska/ddk/keygen/test/KeyPairCreatorTest.java
@@ -13,7 +13,7 @@ import lombok.extern.java.Log;
 import org.junit.Before;
 import org.junit.Test;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 
 import static global.eska.ddk.keygen.utils.Etalon.*;
 import static global.eska.ddk.keygen.utils.HashUtils.sha256;
@@ -60,19 +60,19 @@ public class KeyPairCreatorTest {
 
     @Test
     public void getAddressByPublicKeyNotNullTest() {
-        BigDecimal address = accountCreator.getAddressByPublicKey(PUBLIC_KEY_BYTES);
+        BigInteger address = accountCreator.getAddressByPublicKey(PUBLIC_KEY_BYTES);
         assertNotNull(address);
     }
 
     @Test
     public void getAddressByPublicKeyCorrectAddressTest() {
-        BigDecimal address = accountCreator.getAddressByPublicKey(PUBLIC_KEY_BYTES);
+        BigInteger address = accountCreator.getAddressByPublicKey(PUBLIC_KEY_BYTES);
         assertEquals(ADDRESS, address);
     }
 
     @Test
     public void getAddressByPublicKeyNotNegativeAddressTest() {
-        BigDecimal address = accountCreator.getAddressByPublicKey(PUBLIC_KEY_BYTES);
+        BigInteger address = accountCreator.getAddressByPublicKey(PUBLIC_KEY_BYTES);
         assertTrue(address.signum() > 0);
     }
 }
